Refactor: add index.ts in each app/lib
- add index.ts to import each files
