+++ app/lib/utils/array.ts
@@ -1,11 +1,1 @@
-import { IPicture } from "@/app/lib/types/picture";
-
-export function setPreview(fileList: FileList) {
-  return Array.from(fileList).map((file) =>
-    Object.assign(file, { preview: URL.createObjectURL(file) })
-  );
-}
-
-export function removePreview(pictures: IPicture[]) {
-  pictures.map((picture) => URL.revokeObjectURL(picture.preview));
-}
+import { IPicture } from "@/app/lib/types";
