+++ app/page.tsx
@@ -3,10 +3,9 @@ import { useEffect } from "react";
 import { Divider } from "@/app/ui/components";
 import PictureList from "@/app/ui/picture/List";
 import PictureDetail from "@/app/ui/picture/Detail";
-import { useWidthResizer } from "@/app/lib/hooks/useWidthResizer";
-import { useHeightResizer } from "@/app/lib/hooks/useHeightResizer";
-import { usePictureStore } from "@/app/lib/store/usePictureStore";
-import { removePreview } from "@/app/lib/utils/array";
+import { useHeightResizer, useWidthResizer } from "@/app/lib/hooks";
+import { usePictureStore } from "@/app/lib/store";
+import { removePreview } from "@/app/lib/utils";
 
 export default function Home() {
   const { containerRef, handleMouseDown } = useWidthResizer();
