+++ app/ui/picture/Detail.tsx
@@ -1,7 +1,7 @@
 import { Divider } from "@/app/ui/components";
 import PictureInfo from "@/app/ui/picture/Info";
 import PicturePreview from "@/app/ui/picture/Preview";
-import { useHeightResizer } from "@/app/lib/hooks/useHeightResizer";
+import { useHeightResizer } from "@/app/lib/hooks";
 
 export default function PictureDetail() {
   const { containerRef, handleMouseDown } = useHeightResizer();
