+++ app/ui/picture/ListItem.tsx
@@ -1,6 +1,6 @@
-import { IPicture } from "@/app/lib/types/picture";
+import { IPicture } from "@/app/lib/types";
 import { ListItem } from "@/app/ui/components";
-import { usePictureStore } from "@/app/lib/store/usePictureStore";
+import { usePictureStore } from "@/app/lib/store";
 
 interface PictureListItemProps {
   picture: IPicture;
